requirements.txt
pandas==1.4.2
numpy==1.22.4
fastapi==0.75.1
sqlalchemy==1.4.35
catboost==1.0.6
pydantic==1.9.1
psycopg2-binary==2.9.3
uvicorn==0.16.0
python-dotenv==0.20.0
scikit-learn==1.1.1
nltk==3.7
torch>=1.11.0
transformers>=4.20.0
seaborn>=0.11.2
matplotlib>=3.5.2